+++ trunk/trunk_2319/src/scripts/src/InstanceScripts/Raid_TheEye.cpp
@@ -1183,7 +1183,7 @@ public:
 	
 		else
 		{
-			float val = sRand.rand(100.0f);
+			float val = (float)sRand.rand(100.0f);
 			SpellCast(val);
 		}
 	}
@@ -1299,7 +1299,7 @@ public:
 	
 	void AIUpdate()
 	{
-			float val = sRand.rand(100.0f);
+			float val = (float)sRand.rand(100.0f);
 			SpellCast(val);
 	}
 	
@@ -1429,7 +1429,7 @@ public:
 	
 	void AIUpdate()
 	{
-			float val = sRand.rand(100.0f);
+			float val = (float)sRand.rand(100.0f);
 			SpellCast(val);
 	}
 	
@@ -1551,7 +1551,7 @@ public:
 	
 	void AIUpdate()
 	{
-			float val = sRand.rand(100.0f);
+			float val = (float)sRand.rand(100.0f);
 			SpellCast(val);
 	}
 	
@@ -1760,7 +1760,7 @@ public:
 	
 		else
 		{
-			float val = sRand.rand(100.0f);
+			float val = (float)sRand.rand(100.0f);
 			SpellCast(val);
 		}
 	}
@@ -2039,7 +2039,7 @@ public:
 	
 		else
 		{
-			float val = sRand.rand(100.0f);
+			float val = (float)sRand.rand(100.0f);
 			SpellCast(val);
 		}
 	}
